+++ Filelist
@@ -21,6 +21,7 @@ SRC_ALL =	\
 		src/blob.c \
 		src/blowfish.c \
 		src/buffer.c \
+		src/bufwrite.c \
 		src/change.c \
 		src/channel.c \
 		src/charset.c \
@@ -186,6 +187,7 @@ SRC_ALL =	\
 		src/proto/blob.pro \
 		src/proto/blowfish.pro \
 		src/proto/buffer.pro \
+		src/proto/bufwrite.pro \
 		src/proto/change.pro \
 		src/proto/channel.pro \
 		src/proto/charset.pro \
