+++ src/Make_vms.mms
@@ -2,7 +2,7 @@
 # Makefile for Vim on OpenVMS
 #
 # Maintainer:   Zoltan Arpadffy <arpadffy@polarhome.com>
-# Last change:  2019 Sep 27
+# Last change:  2019 Sep 28
 #
 # This has script been tested on VMS 6.2 to 8.2 on DEC Alpha, VAX and IA64
 # with MMS and MMK
@@ -307,48 +307,210 @@ ALL_CFLAGS_VER = /def=($(MODEL_DEF)$(DEFS)$(DEBUG_DEF)$(PERL_DEF)$(PYTHON_DEF) -
 ALL_LIBS = $(LIBS) $(GUI_LIB_DIR) $(GUI_LIB) \
 	   $(PERL_LIB) $(PYTHON_LIB) $(TCL_LIB) $(RUBY_LIB)
 
-SRC =	arabic.c arglist.c autocmd.c beval.c blob.c blowfish.c buffer.c \
-	change.c charset.c cmdexpand.c cmdhist.c crypt.c crypt_zip.c \
-	debugger.c dict.c diff.c digraph.c drawline.c drawscreen.c edit.c \
-	eval.c evalbuffer.c evalfunc.c \
-	evalvars.c evalwindow.c ex_cmds.c ex_cmds2.c ex_docmd.c ex_eval.c \
+SRC = \
+	arabic.c \
+	arglist.c \
+	autocmd.c \
+	beval.c \
+	blob.c \
+	blowfish.c \
+	buffer.c \
+	bufwrite.c \
+	change.c \
+	charset.c \
+	cmdexpand.c \
+	cmdhist.c \
+	crypt.c \
+	crypt_zip.c \
+	debugger.c \
+	dict.c \
+	diff.c \
+	digraph.c \
+	drawline.c \
+	drawscreen.c \
+	edit.c \
+	eval.c \
+	evalbuffer.c \
+	evalfunc.c \
+	evalvars.c \
+	evalwindow.c \
+	ex_cmds.c \
+	ex_cmds2.c \
+	ex_docmd.c \
+	ex_eval.c \
 	ex_getln.c \
-	if_cscope.c if_xcmdsrv.c fileio.c filepath.c, findfile.c fold.c \
-	getchar.c hardcopy.c hashtab.c highlight.c \
-	indent.c insexpand.c json.c list.c main.c map.c mark.c menu.c mbyte.c \
-	memfile.c memline.c message.c misc1.c misc2.c mouse.c move.c normal.c \
+	fileio.c \
+	filepath.c, \
+	findfile.c \
+	fold.c \
+	getchar.c \
+	hardcopy.c \
+	hashtab.c \
+	highlight.c \
+	if_cscope.c \
+	if_xcmdsrv.c \
+	indent.c \
+	insexpand.c \
+	json.c \
+	list.c \
+	main.c \
+	map.c \
+	mark.c \
+	mbyte.c \
+	memfile.c \
+	memline.c \
+	menu.c \
+	message.c \
+	misc1.c \
+	misc2.c \
+	mouse.c \
+	move.c \
+	normal.c \
 	ops.c \
-	option.c optionstr.c popupmenu.c popupwin.c profiler.c quickfix.c \
-	regexp.c register.c scriptfile.c \
-	search.c session.c sha256.c sign.c spell.c spellfile.c spellsuggest.c \
-	syntax.c tag.c \
-	term.c termlib.c testing.c textprop.c ui.c undo.c usercmd.c \
-	userfunc.c version.c viminfo.c screen.c window.c os_unix.c os_vms.c \
+	option.c \
+	optionstr.c \
+	os_unix.c \
+	os_vms.c \
 	pathdef.c \
-	$(GUI_SRC) $(PERL_SRC) $(PYTHON_SRC) $(TCL_SRC) \
- 	$(RUBY_SRC) $(HANGULIN_SRC) $(MZSCH_SRC) $(XDIFF_SRC)
+	popupmenu.c \
+	popupwin.c \
+	profiler.c \
+	quickfix.c \
+	regexp.c \
+	register.c \
+	screen.c \
+	scriptfile.c \
+	search.c \
+	session.c \
+	sha256.c \
+	sign.c \
+	spell.c \
+	spellfile.c \
+	spellsuggest.c \
+	syntax.c \
+	tag.c \
+	term.c \
+	termlib.c \
+	testing.c \
+	textprop.c \
+	ui.c \
+	undo.c \
+	usercmd.c \
+	userfunc.c \
+	version.c \
+	viminfo.c \
+	window.c \
+	$(GUI_SRC) \
+	$(PERL_SRC) \
+	$(PYTHON_SRC) \
+	$(TCL_SRC) \
+	$(RUBY_SRC) \
+	$(HANGULIN_SRC) \
+	$(MZSCH_SRC) \
+	$(XDIFF_SRC)
 
-OBJ = 	arabic.obj arglist.obj autocmd.obj beval.obj blob.obj blowfish.obj \
-	buffer.obj change.obj charset.obj cmdexpand.obj cmdhist.obj \
-	crypt.obj crypt_zip.obj debugger.obj dict.obj diff.obj digraph.obj \
-	drawline.obj drawscreen.obj edit.obj eval.obj evalbuffer.obj \
-	evalfunc.obj evalvars.obj evalwindow.obj ex_cmds.obj ex_cmds2.obj \
-	ex_docmd.obj ex_eval.obj ex_getln.obj if_cscope.obj if_xcmdsrv.obj \
-	fileio.obj filepath.obj \
-	findfile.obj fold.obj getchar.obj hardcopy.obj hashtab.obj \
-	highlight.obj indent.obj insexpand.obj json.obj list.obj main.obj \
-	map.obj mark.obj menu.obj memfile.obj memline.obj message.obj \
-	misc1.obj misc2.obj mouse.obj move.obj mbyte.obj normal.obj ops.obj \
+OBJ = \
+	arabic.obj \
+	arglist.obj \
+	autocmd.obj \
+	beval.obj \
+	blob.obj \
+	blowfish.obj \
+	buffer.obj \
+	bufwrite.obj \
+	change.obj \
+	charset.obj \
+	cmdexpand.obj \
+	cmdhist.obj \
+	crypt.obj \
+	crypt_zip.obj \
+	debugger.obj \
+	dict.obj \
+	diff.obj \
+	digraph.obj \
+	drawline.obj \
+	drawscreen.obj \
+	edit.obj \
+	eval.obj \
+	evalbuffer.obj \
+	evalfunc.obj \
+	evalvars.obj \
+	evalwindow.obj \
+	ex_cmds.obj \
+	ex_cmds2.obj \
+	ex_docmd.obj \
+	ex_eval.obj \
+	ex_getln.obj \
+	fileio.obj \
+	filepath.obj \
+	findfile.obj \
+	fold.obj \
+	getchar.obj \
+	hardcopy.obj \
+	hashtab.obj \
+	highlight.obj \
+	if_cscope.obj \
+	if_mzsch.obj \
+	if_xcmdsrv.obj \
+	indent.obj \
+	insexpand.obj \
+	json.obj \
+	list.obj \
+	main.obj \
+	map.obj \
+	mark.obj \
+	mbyte.obj \
+	memfile.obj \
+	memline.obj \
+	menu.obj \
+	message.obj \
+	misc1.obj \
+	misc2.obj \
+	mouse.obj \
+	move.obj \
+	normal.obj \
+	ops.obj \
 	option.obj \
-	optionstr.obj popupmenu.obj popupwin.obj profiler.obj quickfix.obj \
-	regexp.obj register.obj scriptfile.obj \
-	search.obj session.obj sha256.obj sign.obj spell.obj spellfile.obj \
-	spellsuggest.obj syntax.obj tag.obj term.obj termlib.obj testing.obj \
+	optionstr.obj \
+	os_unix.obj \
+	os_vms.obj \
+	pathdef.obj \
+	popupmenu.obj \
+	popupwin.obj \
+	profiler.obj \
+	quickfix.obj \
+	regexp.obj \
+	register.obj \
+	screen.obj \
+	scriptfile.obj \
+	search.obj \
+	session.obj \
+	sha256.obj \
+	sign.obj \
+	spell.obj \
+	spellfile.obj \
+	spellsuggest.obj \
+	syntax.obj \
+	tag.obj \
+	term.obj \
+	termlib.obj \
+	testing.obj \
 	textprop.obj \
-	ui.obj undo.obj usercmd.obj userfunc.obj screen.obj version.obj \
-	viminfo.obj window.obj os_unix.obj os_vms.obj pathdef.obj if_mzsch.obj \
-	$(GUI_OBJ) $(PERL_OBJ) $(PYTHON_OBJ) $(TCL_OBJ) \
- 	$(RUBY_OBJ) $(HANGULIN_OBJ) $(MZSCH_OBJ) $(XDIFF_OBJ)
+	ui.obj \
+	undo.obj \
+	usercmd.obj \
+	userfunc.obj \
+	version.obj \
+	viminfo.obj \
+	window.obj \
+	$(GUI_OBJ) \
+	$(PERL_OBJ) \
+	$(PYTHON_OBJ) \
+	$(TCL_OBJ) \
+	$(RUBY_OBJ) \
+	$(HANGULIN_OBJ) \
+	$(MZSCH_OBJ) \
+	$(XDIFF_OBJ)
 
 # Default target is making the executable
 all : [.auto]config.h mmk_compat motif_env gtk_env perl_env python_env tcl_env ruby_env $(TARGET)
@@ -523,6 +685,10 @@ autocmd.obj : autocmd.c vim.h [.auto]config.h feature.h os_unix.h
 blowfish.obj : blowfish.c vim.h [.auto]config.h feature.h os_unix.h
 blob.obj : blob.c vim.h [.auto]config.h feature.h os_unix.h	
 buffer.obj : buffer.c vim.h [.auto]config.h feature.h os_unix.h \
+ ascii.h keymap.h term.h macros.h structs.h regexp.h \
+ gui.h beval.h [.proto]gui_beval.pro option.h ex_cmds.h proto.h \
+ globals.h version.h
+bufwrite.obj : bufwrite.c vim.h [.auto]config.h feature.h os_unix.h \
  ascii.h keymap.h term.h macros.h structs.h regexp.h \
  gui.h beval.h [.proto]gui_beval.pro option.h ex_cmds.h proto.h \
  globals.h version.h
