+++ src/proto.h
@@ -64,6 +64,7 @@ extern int _stricoll(char *a, char *b);
 # include "arglist.pro"
 # include "autocmd.pro"
 # include "buffer.pro"
+# include "bufwrite.pro"
 # include "change.pro"
 # include "charset.pro"
 # include "cmdexpand.pro"
