+++ src/proto/fileio.pro
@@ -1,14 +1,22 @@
 /* fileio.c */
+void filemess(buf_T *buf, char_u *name, char_u *s, int attr);
 int readfile(char_u *fname, char_u *sfname, linenr_T from, linenr_T lines_to_skip, linenr_T lines_to_read, exarg_T *eap, int flags);
 int is_dev_fd_file(char_u *fname);
 int prep_exarg(exarg_T *eap, buf_T *buf);
 void set_file_options(int set_options, exarg_T *eap);
 void set_forced_fenc(exarg_T *eap);
 int check_file_readonly(char_u *fname, int perm);
-int buf_write(buf_T *buf, char_u *fname, char_u *sfname, linenr_T start, linenr_T end, exarg_T *eap, int append, int forceit, int reset_changed, int filtering);
 int vim_fsync(int fd);
+int set_rw_fname(char_u *fname, char_u *sfname);
 void msg_add_fname(buf_T *buf, char_u *fname);
+int msg_add_fileformat(int eol_type);
 void msg_add_lines(int insert_space, long lnum, off_T nchars);
+void msg_add_eol(void);
+int time_differs(long t1, long t2);
+int need_conversion(char_u *fenc);
+int get_fio_flags(char_u *ptr);
+int get_win_fio_flags(char_u *ptr);
+int get_mac_fio_flags(char_u *ptr);
 char_u *shorten_fname1(char_u *full_path);
 char_u *shorten_fname(char_u *full_path, char_u *dir_name);
 void shorten_buf_fname(buf_T *buf, char_u *dirname, int force);
